Refactor RichTextContext into separate file
