+++ src/collections/components/CollectionDetailsPage/form.tsx
@@ -13,10 +13,11 @@ import useHandleFormSubmit from "@saleor/hooks/useHandleFormSubmit";
 import { mapMetadataItemToInput } from "@saleor/utils/maps";
 import getMetadata from "@saleor/utils/metadata/getMetadata";
 import useMetadataChangeTrigger from "@saleor/utils/metadata/useMetadataChangeTrigger";
-import useRichText, {
+import {
   RichTextContext,
   RichTextContextValues
-} from "@saleor/utils/richText/useRichText";
+} from "@saleor/utils/richText/context";
+import useRichText from "@saleor/utils/richText/useRichText";
 import React, { useEffect } from "react";
 
 export interface CollectionUpdateFormData extends MetadataFormData {
