+++ src/shipping/components/ShippingRateInfo/ShippingRateInfo.tsx
@@ -8,7 +8,7 @@ import { commonMessages } from "@saleor/intl";
 import { makeStyles } from "@saleor/macaw-ui";
 import { getFormErrors } from "@saleor/utils/errors";
 import getShippingErrorMessage from "@saleor/utils/errors/shipping";
-import { useRichTextContext } from "@saleor/utils/richText/useRichText";
+import { useRichTextContext } from "@saleor/utils/richText/context";
 import React from "react";
 import { defineMessages, useIntl } from "react-intl";
